#42 renamed the kubernetes-client module to kubernetes
