+++ src/modules/kubernetes/kubernetes.controller.ts
@@ -1,7 +1,7 @@
-import { DeploymentInterceptor } from './../../interceptors/deployment.interceptor';
+import { DeploymentInterceptor } from '../../interceptors/deployment.interceptor';
 import { DeploymentMetricsDto } from './dto/deployment-metrics.dto';
-import { KubernetesClientService } from './kubernetes-client.service';
-import { WorkspaceInterceptor } from './../../interceptors/workspace.interceptor';
+import { KubernetesService } from './kubernetes.service';
+import { WorkspaceInterceptor } from '../../interceptors/workspace.interceptor';
 import { JwtAuthenticationGuard } from '../authentication/guards/jwt-authentication.guard';
 import {
   Controller,
@@ -14,20 +14,18 @@ import {
 @UseGuards(JwtAuthenticationGuard)
 @UseInterceptors(WorkspaceInterceptor, DeploymentInterceptor)
 @Controller('api/workspaces/:workspaceId/deployments/:deploymentId')
-export class KubernetesClientController {
-  constructor(
-    private readonly kubernetesClientService: KubernetesClientService,
-  ) {}
+export class KubernetesController {
+  constructor(private readonly kubernetesService: KubernetesService) {}
 
   @Get('logs')
   findOneLogs(@Param('deploymentId') deploymentId: string): Promise<string> {
-    return this.kubernetesClientService.getPodLogs(deploymentId);
+    return this.kubernetesService.getPodLogs(deploymentId);
   }
 
   @Get('metrics')
   findOneMetrics(
     @Param('deploymentId') deploymentId: string,
   ): Promise<DeploymentMetricsDto> {
-    return this.kubernetesClientService.getPodMetrics(deploymentId);
+    return this.kubernetesService.getPodMetrics(deploymentId);
   }
 }
