+++ src/modules/kubernetes/schemas/pod-condition.schema.ts
