+++ src/modules/kubernetes/schemas/pod-phase.schema.ts
